+++ changelog.md
@@ -9,6 +9,8 @@ Changelog
 * Any path in the frontend can now be overridden from the data/ directory.
 * Allow users to override anything in `data` directory, including `index.html`.
 * Now licensed as BSD 3-clause.
+* If you created your own hw-window elements, they just have been renamed to
+  hw-group. hw-window is a more general window component now.
 
 
 0.1.0 (2026-05-25)